refactor(editor-ui): JSON mapping move more logic to new component
